+++ src/collective/jsonmigrator/blueprints/source_remote.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from base64 import encodestring
 from collective.jsonmigrator import logger
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
@@ -20,6 +19,12 @@ import string
 
 _marker = object()
 MEMOIZE_PROPNAME = "_memojito_"
+
+if six.PY2:
+    # BBB: base64.encodebytes doesn't exist in Python 2
+    from base64 import encodestring as encodebytes
+else:
+    from base64 import encodebytes
 
 if six.PY2:
     # BBB: json.JSONDecodeError doesn't exist in Python 2
@@ -71,7 +76,7 @@ class BasicAuth(six.moves.xmlrpc_client.Transport):
                 "AUTHORIZATION",
                 "Basic %s"
                 % string.replace(
-                    encodestring("%s:%s" % (self.username, self.password)), "\012", ""
+                    encodebytes("%s:%s" % (self.username, self.password)), "\012", ""
                 ),
             )
         h.endheaders()